Refactor medicine package structure
