+++ HealthInMind/app/src/main/java/course/labs/healthinmind/common/dependencyinjection/CompositionRoot.java
@@ -2,12 +2,10 @@ package course.labs.healthinmind.common.dependencyinjection;
 
 import android.app.Application;
 
-import course.labs.healthinmind.medecine.MedicineFactory;
-import course.labs.healthinmind.medecine.MedicineFactoryRoomImpl;
-import course.labs.healthinmind.medecine.MedicineRepository;
-import course.labs.healthinmind.medecine.MedicineRepositoryImpl;
-import course.labs.healthinmind.medecine.MedicineRoomDatabaseProvider;
-import course.labs.healthinmind.medecine.MedicinesLocalProvider;
+import course.labs.healthinmind.medecine.data.abstractions.MedicineRepository;
+import course.labs.healthinmind.medecine.data.roomimplementation.MedicineRepositoryImpl;
+import course.labs.healthinmind.medecine.data.roomimplementation.MedicineRoomDatabaseProvider;
+import course.labs.healthinmind.medecine.data.abstractions.MedicinesLocalProvider;
 
 public class CompositionRoot {
     private Application application;
