+++ HealthInMind/app/src/main/java/course/labs/healthinmind/medecine/data/abstractions/Medicine.java
@@ -1,4 +1,4 @@
-package course.labs.healthinmind.medecine;
+package course.labs.healthinmind.medecine.data.abstractions;
 
 import java.time.LocalTime;
 import java.util.Date;
