+++ HealthInMind/app/src/main/java/course/labs/healthinmind/medecine/data/abstractions/MedicineFactory.java
@@ -1,6 +1,5 @@
-package course.labs.healthinmind.medecine;
+package course.labs.healthinmind.medecine.data.abstractions;
 
-import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Date;
 import java.util.List;
