+++ HealthInMind/app/src/main/java/course/labs/healthinmind/medecine/data/roomimplementation/MedicineImpl.java
@@ -1,4 +1,4 @@
-package course.labs.healthinmind.medecine;
+package course.labs.healthinmind.medecine.data.roomimplementation;
 
 import androidx.room.Embedded;
 import androidx.room.Junction;
@@ -10,6 +10,7 @@ import java.util.Date;
 import java.util.List;
 
 import course.labs.healthinmind.database.MedicineReminderCrossRef;
+import course.labs.healthinmind.medecine.data.abstractions.Medicine;
 import course.labs.healthinmind.reminders.ReminderRoomImpl;
 
 public class MedicineImpl implements Medicine {
