+++ HealthInMind/app/src/main/java/course/labs/healthinmind/medecine/domain/ShowMedicinesUseCase.java
@@ -1,6 +1,9 @@
-package course.labs.healthinmind.medecine;
+package course.labs.healthinmind.medecine.domain;
 
 import java.util.List;
+
+import course.labs.healthinmind.medecine.data.abstractions.Medicine;
+import course.labs.healthinmind.medecine.data.abstractions.MedicineRepository;
 
 public class ShowMedicinesUseCase {
     private MedicineRepository repository;
